copy Installer.desktop only if force_installer=true
Former-commit-id: 415364036cf04278061aff921e10e111ead728ef [formerly a48d708ef20a6f0ee65ee7bf36f696398412096b]
Former-commit-id: 46b5e49df983413c7bf9141677e4f597b9543d3c

+++ mxsnapshot.cpp
@@ -477,9 +477,10 @@ void mxsnapshot::setupEnv()
     if (reset_accounts) {
         system("installed-to-live -b /.bind-root start empty=/home general version-file read-only");
     } else {
-        // copy minstall.desktop to Desktop on all accounts
-        system("echo /home/*/Desktop | xargs -n1 cp /usr/share/applications/minstall.desktop 2>/dev/null");
-        system("chmod +x /home/*/Desktop/minstall.desktop");
+        if (force_installer == true) {  // copy minstall.desktop to Desktop on all accounts
+            system("echo /home/*/Desktop | xargs -n1 cp /usr/share/applications/minstall.desktop 2>/dev/null");
+            system("chmod +x /home/*/Desktop/minstall.desktop");
+        }
         system("installed-to-live -b /.bind-root start bind=/home live-files version-file adjtime read-only");
     }
 }
